+++ Modules/xDFS/DSCResources/MSFT_xDFSNamespaceServerConfiguration/MSFT_xDFSNamespaceServerConfiguration.psm1
@@ -56,15 +56,9 @@ function Get-TargetResource
             $($LocalizedData.GettingNamespaceServerConfigurationMessage)
         ) -join '' )
 
-    <#
-        The ComputerName will always be localhost unless a good reason can be provided to
-        enable it as a parameter.
-    #>
-    $computerName = 'localhost'
-
     # Get the current DFSN Server Configuration
     $serverConfiguration = Get-DfsnServerConfiguration `
-        -ComputerName $computerName `
+        -ComputerName $env:COMPUTERNAME `
         -ErrorAction Stop
 
     # Generate the return object.
@@ -129,15 +123,9 @@ function Set-TargetResource
             $($LocalizedData.SettingNamespaceServerConfigurationMessage)
         ) -join '' )
 
-    <#
-        The ComputerName will always be localhost unless a good reason can be provided to
-        enable it as a parameter.
-    #>
-    $computerName = 'localhost'
-
     # Get the current DFSN Server Configuration
     $serverConfiguration = Get-DfsnServerConfiguration `
-        -ComputerName $computerName `
+        -ComputerName $env:COMPUTERNAME `
         -ErrorAction Stop
 
     # Generate a list of parameters that will need to be changed.
@@ -172,7 +160,7 @@ function Set-TargetResource
     {
         # Update any parameters that were identified as different
         $null = Set-DfsnServerConfiguration `
-            -ComputerName $computerName `
+            -ComputerName $env:COMPUTERNAME `
             @changeParameters `
             -ErrorAction Stop
 
@@ -245,18 +233,12 @@ function Test-TargetResource
             $($LocalizedData.TestingNamespaceServerConfigurationMessage)
         ) -join '' )
 
-    <#
-        The ComputerName will always be localhost unless a good reason can be provided to
-        enable it as a parameter.
-    #>
-    $computerName = 'localhost'
-
     # Flag to signal whether settings are correct
     [System.Boolean] $desiredConfigurationMatch = $true
 
     # Get the current DFSN Server Configuration
     $serverConfiguration = Get-DfsnServerConfiguration `
-        -ComputerName $computerName `
+        -ComputerName $env:COMPUTERNAME `
         -ErrorAction Stop
 
     # Check each parameter